backend/requirements.txt
bidict==0.23.1
blinker==1.9.0
click==8.3.1
dnspython==2.8.0
et_xmlfile==2.0.0
eventlet==0.33.3
Flask==3.0.0
Flask-PyMongo==2.3.0
Flask-SocketIO==5.3.6
Flask-SQLAlchemy==3.1.1
greenlet==3.3.1
h11==0.16.0
itsdangerous==2.2.0
Jinja2==3.1.6
kafka-python-ng==2.2.3
MarkupSafe==3.0.3
marshmallow==3.20.1
numpy==2.4.2
openpyxl==3.1.5
packaging==26.0
pandas==3.0.0
psycopg2-binary==2.9.9
pymongo==4.16.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.0
python-engineio==4.13.1
python-socketio==5.16.1
simple-websocket==1.1.0
six==1.17.0
SQLAlchemy==2.0.46
typing_extensions==4.15.0
Werkzeug==3.1.5
wsproto==1.3.2
